# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Captora Photography CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Unauthorized
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/status:
    patch:
      operationId: updateClientStatus
      tags: [clients]
      summary: Move client to a new pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}/notes:
    get:
      operationId: listClientNotes
      tags: [clients]
      summary: List notes for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientNote"
    post:
      operationId: addClientNote
      tags: [clients]
      summary: Add a note to a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientNoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientNote"

  # SHOOTS
...
  /content:
    get:
      operationId: listContentIdeas
      tags: [content]
      summary: List all content ideas
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of content ideas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentIdea"
    post:
      operationId: createContentIdea
      tags: [content]
      summary: Create a content idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentIdeaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentIdea"
  /content/{id}:
    patch:
      operationId: updateContentIdea
      tags: [content]
      summary: Update a content idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentIdeaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentIdea"
    delete:
      operationId: deleteContentIdea
      tags: [content]
      summary: Delete a content idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthToken:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string
...
    Client:
      type: object
      required: [id, name, phone, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        weddingDate:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        functions:
          type: array
          items:
            type: string
        status:
          type: string
        packageAmount:
          type: ["number", "null"]
        totalPaid:
          type: number
        totalPending:
          type: number
        createdAt:
          type: string
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            notes:
              type: array
              items:
                $ref: "#/components/schemas/ClientNote"
            payments:
              type: array
              items:
                $ref: "#/components/schemas/Payment"
            shoots:
              type: array
              items:
                $ref: "#/components/schemas/Shoot"
            deliverables:
              $ref: "#/components/schemas/Deliverable"
    ClientInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        weddingDate:
          type: string
        venue:
          type: string
        city:
          type: string
        functions:
          type: array
          items:
            type: string
        status:
          type: string
        packageAmount:
          type: number
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        weddingDate:
          type: string
        venue:
          type: string
        city:
          type: string
        functions:
          type: array
          items:
            type: string
        packageAmount:
          type: number
    ClientStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ClientNote:
      type: object
      required: [id, clientId, content, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
    ClientNoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    Shoot:
      type: object
      required: [id, clientId, shootDate, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        shootDate:
          type: string
        shootTime:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        functions:
          type: array
          items:
            type: string
        status:
          type: string
        specialInstructions:
          type: ["string", "null"]
        deliveryNotes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Payment:
      type: object
      required: [id, clientId, amount, paymentDate, installmentType, mode, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        amount:
          type: number
        paymentDate:
          type: string
        installmentType:
          type: string
        mode:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Deliverable:
      type: object
      required: [id, clientId]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        editedPhotos:
          type: boolean
        cinematicHighlight:
          type: boolean
        traditionalVideo:
          type: boolean
        instagramReels:
          type: boolean
        albumOrdered:
          type: boolean
        albumDelivered:
          type: boolean
        rawDataCopied:
          type: boolean
        magazineDelivered:
          type: boolean
        photoFrameDelivered:
          type: boolean
        status:
          type: string
        completedAt:
          type: ["string", "null"]
        updatedAt:
          type: string
...
    ContentIdea:
      type: object
      required: [id, title, type, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        status:
          type: string
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        shootId:
          type: ["integer", "null"]
        createdAt:
          type: string
    ContentIdeaInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        type:
          type: string
        status:
          type: string
        assignedTo:
          type: string
        dueDate:
          type: string
        shootId:
          type: integer
    ContentIdeaUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        status:
          type: string
        assignedTo:
          type: string
        dueDate:
          type: string
        shootId:
          type: integer