# lib/api-zod/src/generated/types/deliverable.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Captora Photography CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Deliverable {
  id: number;
  clientId: number;
  editedPhotos?: boolean;
  cinematicHighlight?: boolean;
  traditionalVideo?: boolean;
  instagramReels?: boolean;
  albumOrdered?: boolean;
  albumDelivered?: boolean;
  rawDataCopied?: boolean;
  magazineDelivered?: boolean;
  photoFrameDelivered?: boolean;
  status?: string;
  /** @nullable */
  completedAt?: string | null;
  updatedAt?: string;
}