# lib/api-zod/src/generated/types/shoot.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Captora Photography CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Shoot {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  shootDate: string;
  /** @nullable */
  shootTime?: string | null;
  /** @nullable */
  venue?: string | null;
  functions?: string[];
  status: string;
  /** @nullable */
  specialInstructions?: string | null;
  /** @nullable */
  deliveryNotes?: string | null;
  createdAt: string;
}